Move OperatorConfig to a separate file. Fix paths.
